// src/common/enums/errorType.enum.ts
export enum ValidationErrorTypeEnum {
  DEFAULT = 100,
  IS_UUID = 101,
  IS_STRING = 102,
  IS_ARRAY = 103,
  IS_EMAIL = 104,
  IS_HEX = 105,
  IS_BOOLEAN = 106,
  IS_OBJECT = 107,
  IS_NUMBER = 108,
  IS_ENUM = 109,
  IS_NOT_EMPTY = 110,
  IS_DEFINED = 111,
  IS_NOT_EMPTY_OBJECT = 112,
  IS_LENGTH = 113,
  MIN = 114,
  MAX = 115,
  ARRAY_MIN_SIZE = 116,
  ARRAY_MAX_SIZE = 117,
  IS_VALID_REGEX = 118,
  IS_TYPE = 119,
  IS_IN = 120,
  IS_UNIQUE = 121,
  IS_STRING_OR_NULL = 122,
  DUPLICATE_VALUE = 123,
  IS_VALID_ROW_ELEMENT = 124,
  WHITE_LIST_VALIDATION = 125, // property should not exist
  IS_NOT_EMPTY_ARRAY = 126,
  IS_DATE = 127,
  IS_PHONE = 128,
  IS_URL = 129,
  NOT_FOUND = 130,
  IS_MONGO_ID = 131,
  IS_NULL = 132,
  IS_VALID_ACCESS_CONDITION_OPERATOR = 133,
  IS_VALID_ACCESS_CONDITION_VALUE = 134,
  IS_AFTER_NOW = 135,
  IS_VALID_VARIABLE_ARRAYS_LENGTH = 136,
  IS_UNIQUE_ARRAY_OF_STRING = 137,
  IS_NOT_BLANK = 138,
  IS_EXISTS_TITLE_OR_CONTENT_IN_OBJECT = 139,
  IS_POSITIVE = 140,
  IS_INT = 141,
  SHOULD_OR_NOT_EXISTS_BY_PROPERTY = 142,
}
// ...
export enum ClassValidatorErrorNameEnum {
  IS_NOT_EMPTY = 'isNotEmpty',
  IS_DEFINED = 'isDefined',
  IS_VALID_REGEX = 'regexValidator',
  IS_VALID_ROW_ELEMENT = 'isValidRowElement',
  IS_STRING = 'isString',
  IS_UUID = 'isUuid',
  IS_ARRAY = 'isArray',
  IS_EMAIL = 'isEmail',
  IS_HEX = 'isHex',
  IS_BOOLEAN = 'isBoolean',
  IS_OBJECT = 'isObject',
  IS_NUMBER = 'isNumber',
  IS_ENUM = 'isEnum',
  IS_LENGTH = 'isLength',
  MIN = 'min',
  MAX = 'max',
  ARRAY_MIN_SIZE = 'arrayMinSize',
  ARRAY_MAX_SIZE = 'arrayMaxSize',
  WHITE_LIST_VALIDATION = 'whitelistValidation',
  IS_NOT_EMPTY_ARRAY = 'arrayNotEmpty',
  IS_TYPE = 'nestedValidation',
  IS_IN = 'isIn',
  IS_DATE = 'isDateString',
  IS_PHONE = 'isPhoneNumber',
  IS_URL = 'isUrl',
  IS_MONGO_ID = 'isMongoId',
  IS_VALID_ACCESS_CONDITION_OPERATOR = 'isValidExperimentConditionsOperator',
  IS_VALID_ACCESS_CONDITION_VALUE = 'isValidExperimentConditionsValue',
  IS_AFTER_NOW = 'isAfterNow',
  IS_VALID_VARIABLE_ARRAYS_LENGTH = 'isValidVariableArraysLength',
  IS_UNIQUE_ARRAY_OF_STRING = 'isUniqueArrayOfString',
  IS_NOT_BLANK = 'isNotBlank',
  IS_EXISTS_TITLE_OR_CONTENT_IN_OBJECT = 'isExistsTitleOrContentInObject',
  IS_POSITIVE = 'isPositive',
  IS_INT = 'isInt',
  SHOULD_OR_NOT_EXISTS_BY_PROPERTY = 'shouldOrNotExistsByProperty',
}